WEBAPP-74: Move reducers to endpoint package
