--- src/endpoints/index.js
@@ -1,13 +1,0 @@
-import LANGUAGE_ENDPOINT from './language'
-import LOCATION_ENDPOINT from './location'
-import PAGE_ENDPOINT from './page'
-import EVENTS_ENDPOINT from './events'
-import DISCLAIMER_ENDPOINT from './disclaimer'
-// todo: refactor, add redux middleware
-export default [
-  LANGUAGE_ENDPOINT,
-  LOCATION_ENDPOINT,
-  PAGE_ENDPOINT,
-  DISCLAIMER_ENDPOINT,
-  EVENTS_ENDPOINT
-]
